Thread-safe counter per class
All of the primitives and simple types now have their own Counter. This
behaviour is much more expectable, since the eval'ed value of one does
not influence the value of any other type.

Furthermore, the Counter is thread-safe now. In the future, we might add
different implementations of Counters (thread-safe, not thread-safe (for
speed) or even thread-local). For now, the thread-safe variant is a sane
default.